+++ src/lib.rs
@@ -319,8 +319,8 @@ pub trait Cached<K, V> {
     /// Attempt to retrieve a cached value with mutable access
     fn cache_get_mut(&mut self, k: &K) -> Option<&mut V>;
 
-    /// Insert a key, value pair
-    fn cache_set(&mut self, k: K, v: V);
+    /// Insert a key, value pair and return the previous value
+    fn cache_set(&mut self, k: K, v: V) -> Option<V>;
 
     /// Get or insert a key, value pair
     fn cache_get_or_set_with<F: FnOnce() -> V>(&mut self, k: K, f: F) -> &mut V;