+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add Cached implementation for std HashMap
 ## Changed
 - trait Cached have a new method cache_get_or_set_with
+- cache_set now return the previous value if any
 ## Removed
 
 ## [0.14.0]