+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [Unreleased]
 ## Added
-- add Cached implementation for std HashMap
 ## Changed
-- trait Cached have a new method cache_get_or_set_with
-- cache_set now return the previous value if any
+## Removed
+
+## [0.15.0]
+## Added
+- add `Cached` implementation for std `HashMap`
+## Changed
+- trait `Cached` has a new method `cache_get_or_set_with`
+- `cache_set` now returns the previous value if any
 ## Removed
 
 ## [0.14.0]